refactor of filestructure
