+++ app.js
@@ -23,7 +23,7 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(cookieParser());
 console.log(__dirname);
-app.use(express.static(__dirname + '/../app'));
+app.use(express.static(__dirname + '/./public'));
 
 app.use(session({
     secret: 'jakester',
