--- app/.buildignore
@@ -1,1 +1,0 @@
-*.coffee
