+++ package.json
@@ -1,39 +1,22 @@
 {
-  "name": "pathlete",
+  "name": "server",
   "version": "0.0.0",
-  "dependencies": {},
-  "devDependencies": {
-    "grunt": "^0.4.1",
-    "grunt-autoprefixer": "^0.7.3",
-    "grunt-concurrent": "^0.5.0",
-    "grunt-contrib-clean": "^0.5.0",
-    "grunt-contrib-concat": "^0.4.0",
-    "grunt-contrib-connect": "^0.7.1",
-    "grunt-contrib-copy": "^0.5.0",
-    "grunt-contrib-cssmin": "^0.9.0",
-    "grunt-contrib-htmlmin": "^0.3.0",
-    "grunt-contrib-imagemin": "^0.7.0",
-    "grunt-contrib-jshint": "^0.10.0",
-    "grunt-contrib-uglify": "^0.4.0",
-    "grunt-contrib-watch": "^0.6.1",
-    "grunt-filerev": "^0.2.1",
-    "grunt-google-cdn": "^0.4.0",
-    "grunt-newer": "^0.7.0",
-    "grunt-ngmin": "^0.0.3",
-    "grunt-svgmin": "^0.4.0",
-    "grunt-usemin": "^2.1.1",
-    "grunt-wiredep": "^1.7.0",
-    "jshint-stylish": "^0.2.0",
-    "load-grunt-tasks": "^0.4.0",
-    "time-grunt": "^0.3.1",
-    "karma-jasmine": "~0.3.5",
-    "grunt-karma": "~0.10.1",
-    "karma-phantomjs-launcher": "~0.1.4"
-  },
-  "engines": {
-    "node": ">=0.10.0"
-  },
+  "private": true,
   "scripts": {
-    "test": "grunt test"
+    "start": "node ./bin/www"
+  },
+  "dependencies": {
+    "body-parser": "~1.10.2",
+    "cookie-parser": "~1.3.3",
+    "debug": "~2.1.1",
+    "express": "~4.11.1",
+    "express-session": "^1.10.1",
+    "firebase": "^2.1.1",
+
+    "jade": "~1.9.1",
+    "morgan": "~1.5.1",
+    "passport": "^0.2.1",
+    "passport-fitbit": "^0.1.2",
+    "serve-favicon": "~2.2.0"
   }
 }
