+++ public/scripts/controllers/main.js
