+++ public/scripts/controllers/toolbar.js
