--- server/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "server",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "start": "node ./bin/www"
-  },
-  "dependencies": {
-    "body-parser": "~1.10.2",
-    "cookie-parser": "~1.3.3",
-    "debug": "~2.1.1",
-    "express": "~4.11.1",
-    "express-session": "^1.10.1",
-    "firebase": "^2.1.1",
-
-    "jade": "~1.9.1",
-    "morgan": "~1.5.1",
-    "passport": "^0.2.1",
-    "passport-fitbit": "^0.1.2",
-    "serve-favicon": "~2.2.0"
-  }
-}
