--- test/spec/controllers/about.js
@@ -1,22 +1,0 @@
-'use strict';
-
-describe('Controller: AboutCtrl', function () {
-
-  // load the controller's module
-  beforeEach(module('pathleteApp'));
-
-  var AboutCtrl,
-    scope;
-
-  // Initialize the controller and a mock scope
-  beforeEach(inject(function ($controller, $rootScope) {
-    scope = $rootScope.$new();
-    AboutCtrl = $controller('AboutCtrl', {
-      $scope: scope
-    });
-  }));
-
-  it('should attach a list of awesomeThings to the scope', function () {
-    expect(scope.awesomeThings.length).toBe(3);
-  });
-});
